requirements.txt
requests>=2.31
python-dotenv>=1.0
supabase>=2.0
pandas>=2.0
pytest>=7.0
